test: added playwright e2e testing

+++ backend/src/services/posts.js
@@ -1,7 +1,8 @@
 import { Post } from '../db/models/post.js'
+import { User } from '../db/models/user.js'
 
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags })
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
   return await post.save()
 }
 
@@ -16,9 +17,12 @@ export async function listAllPosts(options) {
   return await listPosts({}, options)
 }
 
-export async function listPostsByAuthor(author, options) {
-  const data = { author: { $regex: author, $options: 'i' } }
-  return await listPosts(data, options)
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({
+    username: { $regex: `${authorUsername.trim()}`, $options: 'i' },
+  })
+  if (!user) return []
+  return await listPosts({ author: user._id }, options)
 }
 
 export async function listPostsByTag(tags, options) {
@@ -29,14 +33,14 @@ export async function getPostById(postId) {
   return await Post.findById(postId)
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function updatePost(userId, postId, { title, contents, tags }) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, contents, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, contents, tags } },
     { new: true },
   )
 }
 
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
